refactor: move ReAct strategy into its own package
